Problem Solving
A1. Good Matrix Elements

The Smart Beaver from ABBYY got hooked on square matrices. Now he is busy studying an n × n size matrix,
where n is odd. The Smart Beaver considers the following matrix elements good:

Elements of the main diagonal.
Elements of the secondary diagonal.
Elements of the "middle" row — the row which has exactly  rows above it and the same number of rows below it.
Elements of the "middle" column — the column that has exactly  columns to the left of it and the same number of columns to
the right of it.

The figure shows a 5 × 5 matrix. The good elements are marked with green.
Help the Smart Beaver count the sum of good elements of the given matrix.

Input
The first line of input data contains a single odd integer n. Each of the next n lines contains n integers
aij (0 ≤ aij ≤ 100) separated by single spaces — the elements of the given matrix.

The input limitations for getting 30 points are:

1 ≤ n ≤ 5
The input limitations for getting 100 points are:

1 ≤ n ≤ 101
Output
Print a single integer — the sum of good matrix elements.